+++ demo/lit-app/custom-elements-manifest.config.mjs
@@ -1,7 +1,10 @@
 import { customElementVsCodePlugin } from "custom-element-vs-code-integration";
 import { customElementJetBrainsPlugin } from "custom-element-jet-brains-integration";
 import { customElementSolidJsPlugin } from "custom-element-solidjs-integration";
+import { customElementJsxPlugin } from "custom-element-jsx-integration";
 import { getTsProgram, expandTypesPlugin } from "cem-plugin-expanded-types";
+import { customElementReactWrapperPlugin } from "custom-element-react-wrappers";
+import { customElementVuejsPlugin } from "custom-element-vuejs-integration";
 
 export default {
   /** Globs to analyze */
@@ -33,7 +36,37 @@ export default {
     customElementJetBrainsPlugin({ typesSrc: "expandedType" }),
     customElementSolidJsPlugin({
       // globalTypePath: "./types"
-      componentTypePath: (name, tag) => `./types/${tag}/${name}.d.ts`
-    })
+      componentTypePath: (name, tag) => `./dist/${tag}/${name}.d.ts`,
+    }),
+    customElementJsxPlugin({
+      componentTypePath: (name, tag) => `./dist/${tag}/${name}.d.ts`,
+    }),
+    customElementReactWrapperPlugin({
+      outdir: "./dist/react",
+      modulePath: () => `../index.js`,
+      globalEvents: [
+        {
+          event: "onChange",
+          description: "Fired when the value of an input element changes.",
+          type: "React.ChangeEventHandler",
+        },
+        {
+          event: "onSubmit",
+          description: "Fired when a form is submitted.",
+          type: "React.FormEventHandler",
+        },
+        {
+          event: "onFocus",
+          description: "Triggered when an element receives focus.",
+          type: "React.FocusEventHandler",
+        },
+      ],
+      attributeMapping: {
+        "for": "_for",
+      }
+    }),
+    customElementVuejsPlugin({
+      componentTypePath: (name, tag) => `./dist/${tag}/${name}.d.ts`,
+    }),
   ],
 };