+++ demo/preact-app/scoped-components.d.ts
@@ -1,0 +1,8 @@
+import type { CustomElements, ScopedElements } from "lit-app/custom-element-jsx";
+
+declare module "preact" {
+  namespace JSX {
+    interface IntrinsicElements
+      extends CustomElements {}
+  }
+}