+++ demo/vue-app/vite.config.ts
@@ -1,0 +1,17 @@
+import { defineConfig } from "vite";
+import vue from "@vitejs/plugin-vue";
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [
+    // @ts-ignore
+    vue({
+      template: {
+        compilerOptions: {
+          // treat all tags with a dash as custom elements
+          isCustomElement: (tag) => tag.includes("-"),
+        },
+      },
+    }),
+  ],
+});