+++ packages/jsx-integration/src/index.ts
@@ -1,0 +1,2 @@
+export { generateJsxTypes } from "./type-generator";
+export { customElementJsxPlugin } from "./cem-analyzer-plugin";