+++ packages/solidjs-integration/src/type-generator.ts
@@ -1,12 +1,12 @@
 import {
   Component,
-  EXCLUDED_TYPES,
   getComponentDetailsTemplate,
   getComponentProperties,
   getComponents,
+  getCustomEventTypes,
   getMemberDescription,
 } from "../../../tools/cem-utils";
-import { logBlue, saveFile } from "../../../tools/integrations";
+import { createOutDir, logBlue, saveFile } from "../../../tools/integrations";
 import { Options } from "./types";
 
 export function generateSolidJsTypes(manifest: any, options: Options) {
@@ -16,6 +16,7 @@ export function generateSolidJsTypes(manifest: any, options: Options) {
     (x) => x.tagName
   );
   const template = getTypeTemplate(components, options);
+  createOutDir(options.outdir!);
   const outputPath = saveFile(
     options.outdir!,
     options.fileName!,
@@ -36,27 +37,6 @@ function getOptions(options: Options) {
   return options;
 }
 
-function getEventTypes(component: Component, componentNames: string[]) {
-  const types = component.events
-    ?.map((e) => {
-      const eventType = e.type?.text
-        .replace("[]", "")
-        .replace(" | undefined", "");
-      return eventType &&
-        !EXCLUDED_TYPES.includes(eventType) &&
-        !componentNames.includes(eventType) &&
-        !eventType.includes("<") &&
-        !eventType.includes(`{`) &&
-        !eventType.includes("'") &&
-        !eventType.includes(`"`)
-        ? eventType
-        : undefined;
-    })
-    .filter((e) => e !== undefined && !e?.startsWith("HTML"));
-
-  return types?.length ? [...new Set(types)].join(", ") : undefined;
-}
-
 function getTypeTemplate(components: Component[], options: Options) {
   const componentNames = components
     .filter((x) => x.customElement)
@@ -64,7 +44,7 @@ function getTypeTemplate(components: Component[], options: Options) {
   const componentImportStatements =
     typeof options.componentTypePath === "function"
       ? components.map((c) => {
-          const types = getEventTypes(c, componentNames);
+          const types = getCustomEventTypes(c, componentNames);
           return `import type { ${c.name} ${
             types ? `, ${types}` : ""
           } } from "${options.componentTypePath?.(c.name, c.tagName)}";`;
@@ -77,7 +57,7 @@ ${
   options.globalTypePath
     ? `import type { ${components
         .map((c) => {
-          const types = getEventTypes(c, componentNames);
+          const types = getCustomEventTypes(c, componentNames);
           return c.name + (types ? `, ${types}` : "");
         })
         .join(", ")} } from "${options.globalTypePath}";`