+++ packages/expanded-types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.1.4
+
+- Updated to prevent errors when object types have comments
+- Add optional chaining when evaluating types in case they return `null` or undefined
+
+## 1.1.3
+
+- Added check for missing value
+
+## 1.1.2
+
+- Fixed type imports with file extensions
+- Fixed support for combined union types
+
+## 1.1.1
+
+- Add optional chaining to prevent errors when types are `undefined`.
+
 ## 1.1.0
 
 - Update TypeScript configuration to support jsDoc types